adapter: Drop "view not found" log to debug
This log message is hit every time we prepare a non-migrated query for
the first time, since the adapter always asks the server for the view
for a prepared query in case that query was migrated by another adapter.
There's nothing a user can do about this in that case - and in general
this log message is *very* noisy and not very actionable. Let's drop it
down to debug.

Change-Id: I432811936a91857e14ed83aa0a9639bb94fc231d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6117
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>