adapter: Don't warn log on failed to parse
If we fail to migrate a query due to it failing to parse, don't warn log
at all. This happens all the time in real-world applications, both for
invalid queries and (more significantly) queries that we just don't know
how to parse, and there's nothing a user can do about it.

Change-Id: If92f9da0ddc5237bd4e7bd8cab80342ee49fc655