adapter: LazyUpstream.is_connected() should connect
`SHOW READYSET STATUS` is displays the status of the current connection,
inside the adapter, as it connects to the upstream database. This was
inadvertently broken when we introduced LazyUpstream as it intentionally
delays connecting. Thus, when a new psql session immediately executes
`SHOW READYSET STATUS`, it would display "Database Connection:
Unreachable", even though it had never attempted to connect to
the upstream.

This patch changes `LazyUpstream.is_connected()` to actually connect
if not already connected, and defer to inner impls to test the
connection.

Fixes: REA-3541
Change-Id: Ia60b4c5fe8954be189b37498b5e4922e6e108e5c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6120
Tested-by: Buildkite CI
Reviewed-by: Tamas Juhasz <tamas@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>