readyset-psql: Improve upstream connectivity check
`UpstreamDatabase.is_connected()` currently just calls
`tokio_postgres::Client::is_closed()`. That function simply checks
that it's inner `mpsc::UnboundedSender` is not closed. That's not
a sufficient check as if the database is down, we don't actually
check that it is available, we only look at that sender. The sender
won't be shutdown by some background task (ala keep-alives). Thus,
we actually need to issue a probing query to the database.

This patch switches to issuing a simple `select 1` query to the
upstream PG to make sure we can connect.

Change-Id: I977e2d11c7cc6f4bd6b9cdcff3a8e6ecc696363f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6121
Tested-by: Buildkite CI
Reviewed-by: Tamas Juhasz <tamas@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>