adapter: Don't warn if we fail to parse
Similar to 9dc4b3db1e0717785e2d2fd52b0fd5dc8d2d2331 but in a different
place, suppress warn! logs if the reason for the query failure is us
failing to parse. We purposely don't parse unsupported queries and this
is very common to happen and not worth attention currently.

Change-Id: Icde3812ff901af5ea4da3b0ff8d02b6a9e4971ea