server: Allow generating graphviz for a single query
Allow limiting the graphviz generated by the /graphviz RPC endpoint to
only those nodes that are part of a single query. Under the hood, this
means only those nodes reachable via *incoming* edges from the reader
node of that query.

Since this is the second way that graphviz can be configured via
arguments to controller RPCs, and I want to avoid an explosion in
different RPC endpoints, I've also gotten rid of the /simple_graphviz
RPC endpoint in favor of a new GraphvizOptions struct passed to the
`graphviz()` method.

I have, however, exposed this through a (non-bincode!) GET
/graph/query_name controller HTTP endpoint, which can be useful on the
command-line.

Change-Id: Iec676b6c86ceebc4842ec15d9f2b94f5efa930e9