readyset-mysql: Add gauge for upstream connections
To reduce logging noise, lower the logging level of "Establishing
connection" (to the upstream) to debug, and add a gauge to measure
the current upstream connection count. The gauge can be scraped
by prometheus for better observability.

`UpstreamDatabase.reset()` needed to be rewritten as it conflicted
with the new `Drop` implementation. More importantly, `reset()` was
not connecting in the same way that `connect_inner()` is, and we
dropped several connection features (RYW, ssl certificates). Thus,
readyset-mysql's `reset()` function was rewritten to mirror the one in
readyset-psql.

Fixes: REA-3548
Change-Id: I82e2e47779731497e4c9f34fcfd9e8ef850d7a7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6133
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>