buildkite: Add retry on agent failure to jobs
AWS spot instances are subject to reclamation, which causes
buildkite-agent to terminate with code 255.  This adds automatic retry
of jobs if it looks like the job failed because it was on a spot
instance that got yanked back.

Buildkite does not appear to support global retry settings at this time,
so we have to do it on a per job basis.  This update does not address
all pipelines, but it hits the main ones that are likely to be causing
immediate build pain.

Change-Id: I3d7a38f61fbd1d7bad0610f9f2495647ebf6bfa0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6137
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>