readyset-adapter: Add capacity to QueryStatusCache
We are currently seeing proxied queries blow up the QueryStatusCache. As
a first step to mitigating this, allow a capacity to be set after which
we clear the entries in the cache. This will be followed up with a bloom
filter to reduce the impact of the clear as well as eventually some sort
of LRU eviction.

Change-Id: If66a3fc17c0c0c9749df7941bec518838f4821a7