readyset-client: Remove Buffer layer from Controller
The existence of this layer is mostly pretty ancient history at this
point, but my *guess* is that this was added because an early version of
the controller was implemented using a custom TCP wire protocol which
needed to serialize requests over the same socket, requiring us to wrap
it in a Buffer layer to allow ReadySetHandle to impl Clone. Now that
we're using HTTP for controller requests (at *least* as of
020f4b058 (Overhaul noria-server's coordination and RPC system,
2021-05-04), but maybe since before even that), this isn't necessary -
and having this here means that we're accidentally *serializing*
controller requests that go to clones of the same `ReadySetHandle`! We
think that last bit is causing some pretty major performance regressions
when lots of connections to the same adapter make lots of different
queries which all make controller RPCs.

To try to make controller requests from the same adapter (all of which
use clones of the same ReadySetHandle, under the hood) actually
parallel, this commit removes the Buffer layer from the `ReadySetHandle`
entirely, instead opting to just derive Clone for Controller and
RawController directly. I haven't done a perf test on this yet, but it
seems likely to be good.

Change-Id: I10cec4f44807e3693e2acae6400fd4b0fb194f1d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6142
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI