readyset-util: Generalize out a global/local cache
The behavior implemented by noria_connector::ViewCache where we have a
single global map which we create local cached versions of for more
efficient memoization is generally useful, and I'm going to want to use
it a second time (to provide memoization of lookups of `View`s in the
server). In preparation for that, this commit generalizes it out into a
new `readyset_util::global_local` module, with methods implemented for
everything we currently use ViewCache for.

Change-Id: I91fe6803e8dbcce072a0fa46a75f9b4cc2ce0dc7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6143
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>