adapter: Don't lookup the same view twice
Use the new `global_local` module in readyset-util for Views in addition
to the map from select statement to view name, to avoid making
controller RPCs to lookup the same View more than once. For many
workloads this ought to save *significantly* on controller RPCs, since
this allows us to only ever look up the View for a particular query
once.

Change-Id: I62e20be7f31a4be148abe31ccfd9b17f989b538b