replicators: Handle EOF from mysql cleanly
If we kill a mysql docker image that readyset is connected to, we
receive an EOF packet from mysql, which we previously weren't handling
cleanly and panicking. This commit logs a warning if we see EOF instead,
and rather than panicking it returns an error and goes into a retry loop
that can recover if the mysql server comes back up.

Refs: REA-3571
Release-Note-Core: Fixed an issue where ReadySet would panic if it
  received an EOF packet from a mysql upstream that had gone away.
Change-Id: I77a8f392aee0e9d47d3c6c1b7e8785d1f43d1eaf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6147
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>