readyset-psql: Adjust types::citext test assertion
The eventaully! assertion in the types::citext test was doing both the
prepare and the query in the `eventually!` setup. This was a race
condition and citext has been historically flakey. Now it is more flakey
because the prepare is faster thanks to 1fd7fa2a. We can fix this by
moving the prepare outside of the `eventually!` so that only the query
is polled for consistency.

Change-Id: I68ae38b8428d89f82adb499f57a34bb2678f3034