mysql-srv: Add CONNECT_WITH_DB capability.
Suppose the client replies to an initial handshake packet with
CapabilityFlags::CLIENT_CONNECT_WITH_DB set, RS attempts to parse the
database. The problem relies on this flag being set if the client wants
to connect to a specific database. Still, the database name is only
sent as part of the initial handshake packet if the server initially
informs it supports this capability via the CONNECT_WITH_DB flag.
In this case, we might end up reading the next field in the handshake.