adapter: Don't send pending queries to RS
Many applications, such as Rails, have a lifecycle that looks like:

  1. get a request
  2. make a new connection to DB
  3. prepare statement
  4. execute statement
  5. disconnect from DB

In addition, if in Rails the `db_prepared_statements` flag is set to
`false`, the app will *inline* all values for query parameters (but
still use the extended query protocol, confusingly). This can mean
that *every* query sent to us by an application is a "new" query
according to the query status cache, meaning it'd get the status of
Pending - and currently, ReadySet always tries to prepare queries with a
Pending status in the query-status-cache against ReadySet.

What this means when all taken together is that Discourse, a Rails
application which sets `db_prepared_statements` to `false`,
sends *thousands* of requests to the `view_builder` controller RPC when
run under load - basically one per proxied query! This significantly
impacts performance of these workloads as it creates a significant
amount of load on the controller.

Fixes: REA-3533
Change-Id: Id28e3604f1d3605cba03d77af6d1af5e2c0184ad