sql: Filter lhs NULLs from WHERE [NOT] IN (subquery)
lhs IN (subquery) and lhs NOT IN (subquery) both have NULL as the value
if the lhs is NULL, even if the expr would otherwise be true - this is
most notable for NOT IN, where the expr is NULL and therefore the row is
omitted even if the NULL value is not in the rhs subquery at all.

This was previously not implemented correctly for our implementation of
NOT IN - we'd project out a row for the lhs, then keep it in the filter
due to the rhs mark column being NULL. To correct that, this commit adds
a filter node to the lhs of the join to omit any rows where the lhs expr
evaluates to null before joining.

Change-Id: Ia617ce989adb82d557aeaa63ffd5f483c85261e7