sql: Correctly handle NULLs in projected IN (subquery)
The expr `lhs IN (subquery)` evaluates to NULL if the lhs is NULL,
regardless of whether the rhs subquery contains that NULL value. As with
IN (subquery) in the WHERE clause, we were incorrectly handling NULL
values in the lhs in the SELECT list too. This commit fixes that, by
extending the final projected expr for those values to explicitly check
for NULL values in the lhs and make the final value NULL if it is, using
a CASE WHEN expr.

Change-Id: I5b64df995d39bedcd3712ade69b79a6b21c5d44d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6158
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>