readyset-adapter: Track live connections in a SkipSet
Pass around an Arc<SkipSet<SocketAddr>> to each readyset-adapter
Backend, which tracks the set of active connections to the adapter. Each
Backend adds its own addr to the set on startup, and removes itself from
the set on Drop.

This is currently not read from, but will be used to provide a command
to list the currently active connections to the adapter

Change-Id: I383889ed8eb633a1445c38bfcd96fff6e44905fb