adapter: Remove duplicate process_query invocation
In the ad-hoc select path, we were running process_query() twice when we
could re-use the results from the first call later on.

This change also changed the behavior of the ViewCache in that we are
using the rewritten adhoc query for `get_view()` rather than the
pre-rewritten query. This means that the verify_no_placeholder()
function is no longer valid and was thus removed.

Change-Id: I47dbf6c796d4d298a01ea91cf200c30eafa57f06
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6165
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>