adapter: Add id_to_status to QueryStatusCache
This commit continues refactoring the QueryStatusCache by adding a
direct map from QueryId to QueryStatus, which can be used to make query
routing decisions. This commit's main focus is the addition of the extra
mapping as well as changes to the existing QueryStatusKey with_status
and with_mut_status to ensure that this is aligned with the
persistent_handle's view of the status.

The overall vision of this layer is for it to be a lighter-weight, in-memory
counterpart to the persistent state we store that can be rebuilt after a
restart.

This commit also uses this new map to return the status in
query_migration_status.

Change-Id: I7731dc8fc3362484f1c8bb7af8462fbe85ae0e8f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6171
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI