adapter: Limit the number of stored queries in QSC
This is a temporary measure in order to limit the memory usage of the
query status cache by changing the persistent handle (which still isn't
actually persistent yet, but will be) to use an LRUCache as its
underlying store with a limit of 100_000. The id_to_status map is able
to "remember" the query status for more than this number of queries, but
we will only keep the full query for 100_000 queries.

The main goal of this is to remove this as a variable in our locust
testing before the fully persistent refactor can be finished.

Change-Id: I9d39590c3e797db4c0ad40244ad47fce32cb0e51