adapter: Store all prepared statement info in Backend
Previously, information about prepared statements within the adapter was
spread between the `prepared_statements` field in `Backend`,  and the
`PreparedStatementCache` field in `NoriaConnector`. In addition, on the
`NoriaConnector` side the ID of the prepared statement was buried deep
within the various variants of the  `PrepareResult` enum. These two
facts made it difficult to implement deletion of prepared statements,
since the relevant information was spread across multiple places.

In preparation of implementing deletion of prepared statements, this
commit moves *all* information necessary to execute a prepared statement
up to the `prepared_statements` field of `Backend`, including all of the
information from the `noria_connector::PreparedStatement` enum, which is
now gone.

There's a lot of additional miscellanious simplification and improvement
of naming here as well, most of which is just opportunistic.

Change-Id: I3cd269e0ec837406039067b92b19a3eeab5c0078