adapter: Store prepared statements in a Slab
Store prepared statements in the adapter Backend in a Slab rather than
in a Vec, to allow for eventual efficient removal and reuse of space for
information about prepared statements. While this involves moving the
generation of the ID up to insertion time, other than that not much is
changing here.

An earlier iteration of this (I0e393fd87d166a48f073287130d610f8afabf7c6)
used a SlotMap instead, but I decided to go with Slab because it's
difficult to convince a SlotMap to use u32-compatible keys, and the
extra behavior of globally unique IDs that SlotMap provides is probably
not necessary for our use case.

Change-Id: Ic05ca2be700f633b315744c75b4647fc4838c64d