adapter: Replace client command's latencies with counts
SHOW PROXIED QUERIES and SHOW CACHES were recently updated to output
latency percentiles of the queries executed. As we are piggy-
backing off our prometheus metrics for capturing and in-memory
storage, we are bound byt the implementation details of prometheus
histogramws (Distributions). Those Distributions only report data
for the last 60 seconds, by default. The impact to users is if
they execxute a query, dump SHOW CACHES, and see some latency
information, if they simply wait > 60 seconds and issue SHOW
CACHES again, the latencies get "reset to zero".

As we are using this latency information in SHOW CACHES as more
of a Day 1 experience (getting a user to try out and experience
ReadySet), the current "resetting" is a bit confusing. Hence,
this patch replaces the latency percentiles with a simple incrementing
counter. This shows enough information to users to indicate that
ReadySet is indeed working and handling queries.

Fixes: REA-3540
Release-Note-Core: Replace latency histograms in SHOW PROXIED QUERIES
  and SHOW CACHES with a simple counter.
Change-Id: Ic9623ba08f1503491ed283766a0492b2b6e58e37