adapter: Add a remove_statement method to UpstreamDatabase
Add a method to the UpstreamDatabase trait, implemented for both psql
and mysql, that removes a statement from the upstream database. On
postgres, this is done via the Drop impl for StatementInner, but on
MySQL this needs to be done explicitly by calling conn.close(stmt).

This is unused as of this commit, but next up will be used by the client
library shim to fully implement closing of prepared statements

Change-Id: I07def8c5b129eaac6af537c7ce3fb00521507108
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6184
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>