adapter: Use LruCache for parsed query cache
We were storing the full query string as well as the full parsed query
string for every unique parsed query we were seeing. This is quite
expensive if we are constantly seeing new queries.

This commit makes this cache an LRU cache with a capacity (for now) of
10k entries.

Also included in this commit is a benchmark used to decide if we should
use an LruCache<QueryId, SqlQuery> instead of the current version, which
the benchmarks says we should not as it is slower. The extra memory
usage is worth it since this is in the hot path.

Change-Id: I1dfdd629aa933865c852c4d7d04ce4140c0a4040
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6197
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>