adapter: Clean up command line arguments
There are a large number of cli agruments/ev vars we expose, many of
which are only useful for developers of ReadySet and tests. In order
to improve the experience of users, we're hiding many of those options
from the cli help command (the functionality is still available,
of course).

Fixes: REA-3585
Release-Core-Note: Hiding many developer-level flags
  from the cli help output.
Change-Id: Ifcde6b5e7b87646f7f6502d2f32ad4f3a72b63a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6202
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>