adapter: Rename db-dir to storage-dir
storage-dir is arguably a more generic, and more apt, flag name
than db-dir, to indentify the directory where readyset should
store it's local files. The code changes are near trivial, but much
configuration will need to be updated; that will come in future
commits on this CL.

Additionally, treewide change the configration cli flag from
`--db-dir` to `--storage-dir`, as well as change the env var
`DB_DIR` to `STORAGE_DIR`.

Fixes: REA-3585
Change-Id: I8f19c52a74e042a753d7730e8b93f1cd39d5f335
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6203
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>