dataflow: Simplify domain channel code
This commit makes concrete all of the abstract domain channel code. It
seems like that code was written with the assumption that it would be
used broadly for many different purposes. In practice, we only use it
for domain communication, so we are able to remove most of the type
parameters to greatly simplify the types.

This change has the added benefit that we no longer need to `Box` domain
packets when sending them.

Change-Id: I163b31b7a102fc719956c99c25777f290ad7ff70
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6212
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>