client+df: Move domain channel to df crate
The code we use for inter-domain communication is currently very
general, but in practice, it's only used with one set of concrete types.
To set the stage for us to remove the type parameters from these types,
this commit moves all of the domain channel boilerplate from the
readyset-client crate to the readyset-dataflow crate. Since all of the
dataflow types (e.g. `Packet`) are defined in readyset-dataflow, it
would not have been possible to keep the domain channel code in
readyset-client and import the necessary types to make the channel code
concrete, since all of the dataflow types (e.g. `Packet`) are defined in
readyset-dataflow, and readyset-dataflow depends on readyset-client --
this would have created a circular dependency between readyset-client
and readyset-dataflow.

Change-Id: I2ea42400a1163bbaf84d2f8eb3b0a0f92bb29647
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6211
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>