dataflow: Add domain channel types
This commit adds `DomainSender` and `DomainReceiver` types that wrap the
tokio mpsc sender and receiver types. This will allow us to add, in a
future commit, metrics that track the number of `Packet`s queued for
each domain: these wrapper types will wrap a domain metrics handle and
will allow us to emit a metric in centralized locations
(`DomainSender::send` and `DomainReceiver::recv`), as opposed to having
to emit a metric everywhere we invoke `UnboundedSender::send`.

Change-Id: Ia68fa8a1acd54fbc61f754f965550bf87941da36
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6214
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>