adapter: Add metric for /metrics payload size
We saw that the query log/query log ad hoc can cause the /metrics
payload to grow, so this commit adds an easy way to determine the
current payload size of that endpoint. It also creates an extensible
struct that can be re-used for other http endpoints that may have
interesting metrics to report.

Change-Id: Ia5657ebbb2f471a3d1edf74b087b73bf081508b2