readyset-data: Fix citext coercion
We were encountering flakiness in our `readyset_psql::types::citext`
test where, seemingly, an insert was never making it to the reader node.
This ended up being caused by a race between the first read in the
reader node and the aforementioned insert (though the underlying problem
was caused by a mishandling of citext types). The test consists of a write
to a particular key and then a read of that same key, where the column
type of the key is citext. If the write happens first:

- The base table processes the insert
- The write is propagated to the reader with a key of type text with a
  utf8 collation (which is incorrect since we are using citext!)
- The reader throws away the write because there is a hole at that key
- The subequent read uses a key of type text with a *citext* collation,
  misses, triggers an upquery, and returns the correct results

Note that the type of the key read by the query is *different* than the
key in the `Record` propagated to the reader -- though the above test
passes, this difference in types is what causes this test to fail if we
reorder the read and write like so:

- We issue a read using a key of type text with a citext collation,
  which misses, triggers an upquery, and causes the hole to be filled
  with an empty result
- The base table processes the insert
- The write is propagated to the reader with a key of type text with a
  utf8 collation *which is a distinct value from the key we filled the
  hole with since the collation doesn't match*. This mismatch causes the
  reader to throw the write away since it finds a hole
- We attempt to read from the correct key (the one with citext collation),
  but we don't trigger an upquery because we filled the hole with an
  empty result above. The key point here is that the keys of the write
  and read are distinct values that are actually different slots in the
  index

Because citext is a custom type (just one that happens to be installed
by a Postgres extension), we need to coerce it the same way we coerce
other custom types when the base table receives a citext row in a table
operation; if we don't, we will just treat the value as text with a utf8
collation. This commit fixes the issue by properly coercing writes with
citext values.

Fixes: REA-3062
Change-Id: I5b8167668e0d1f1b5cd682833f7a6696f7070f7d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6222
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>