client-metrics: Add QueryLogMode enum
In the adapter we currently have two flags to control query log
recording (capturing query metrics): --query-log and --query-log-ad-hoc.
These are fairly large hammers and in a degenerate case, can create
enormous payloads to prometheus (the worst we happened to see was >
5 GiB!). Hence, we need to be able to fine tune what is recorded, but
adding in more cli flags is insufficient, and the combinations of flags
will be confusing.

This commit introduces an enum that can replace the cli flags, be simple
enough to use, and be extensible for further refinement. It supports
the existing functionality ("is this enabled", "include all ad-hoc queries"),
and adds the ability to 1) record only cached query information, and 2)
reduce verbosity of histograms by not adding a tag with the full SQL query
text. All of this is still available in the "All" mode (i.e. expert mode),
and is primarily intended for readyset engineers to debug issues.

Fixes: REA-3592
Change-Id: I7725ef8ce43a269dee7e17717d49be32d63829aa