adapter: replace query-logs flags with QueryLogMode
This commit removes `--query-log` and `--query-log-ad-hoc`, and replaces
them with the new QueryLogMode enum. The enum value is specified by
the operator via a new cli flag, `--query-log-mode`. The enum is then
plumbed through the adapter.

The `QueryLogger` will now restrict adding the `query` tag (which
contains the SQL query text) to histograms only when the `All` mode
 is selected. Similarly, query parse histograms are only recorded
 in `All` mode, as well. `Cached` mode will only record latencies from
 queries that were serviced by ReadySet (i.e., they were cached).

Also, this commit cleans up the existing uses of the cli flags and
env vars with approporiate conversion to the new `--quer-log-mode`
and `QUERY_LOG_MODE`.

Fixes: REA-3592
Release-Core-Note: `--query-log` and `--query-log-ad-hoc` are replaced
  by `--query-log-mode`.
Change-Id: I6da3f15d77c66d4ed9e1609b1d1915658b4aba30
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6233
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>