server: Add config flag for allowing straddled joins
Add a config flag for allowing queries containing straddled joins. These
queries can sometimes be slow, so we're adding a config flag, defaulting
to false, to allow migrating them. If the flag is set, the query returns
an unsupported error during materialization planning (which is the only
time we actually know a join is straddled).

Release-Note-Core: Remove support for straddled joins (joins with
  partial keys which trace to each side of the join), which can
  sometimes be slow to execute.
Change-Id: I62e87a10bda2d9fb6b5eeb177a2a10f5fd3a7ec3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6239
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Ethan Donowitz <ethan@readyset.io>