alloc-metrics: Record memory allocator metrics
As we want observability into the jemalloc allocator, this commit
introduces a new crate, `readyset-alloc-metrics`, that fetches the
current jemalloc state via `readyset-alloc`, and publishes those
as metrics (gauges) to our prometheus handler. As the values are
gauges, we need to sample the jemalloc stats periodically.

This new crate is enabled in both adapter and server.

Change-Id: If428829f3f3015ee8d40fdae4655161a8923ba88
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6246
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>