mir: Don't try to fuse nodes with multiple descendants
When we try to fuse nodes, we have an invariant that they must have a
single descendant. The previous logic was checking for this single child
only after calling `remove_node`, but `remove_node` itself was erroring
out in this case. This commit adds a check before that to prevent these
errors.

Change-Id: Ia2ecc124f16da8a325e56cbe04148e445b98744f