adapter: Fix ignored shutdown signal
There was a retry loop to load the schema search path in the adapter
that had ownership of a shutdown receiver but was not listening for a
signal. This caused the shutdown sender to time out as it waited in vain
for that task to drop its shutdown receiver. This commit fixes the issue
by adding a call to `tokio::select!` that correctly listens for the
shutdown signal and returns early if one is received.

Fixes: REA-3575
Release-Note-Core: Fixed an issue where ReadySet failed to shut down in
  certain situations
Change-Id: I246abe106e2698582c7b5a34f54413bd931506e6