replicators: Resnapshot if slot deleted
Previously, when the replicator started up, it expected a replication
slot to exist if we weren't resnapshotting and if every base table had a
replication offset. If the slot did *not* exist, the replicator would
enter an infinite retry loop. This commit fixes the issue by checking if
the slot exists and, if it doesn't, creating a new slot and performing a
full resnapshot.

Fixes: REA-3503
Release-Note-Core: Fixed an issue where ReadySet didn't start up
  correctly if its replication slot was dropped while it was down
Change-Id: Ibead02b3d0986719c5ba687ea447b41a8da0370e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6256
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>