nom-sql: Add support for DEALLOCATE
Client apps can send a DEALLOCATE statement to indicate that either
all or a single prepared statement should be freed. This commit
adds support to nom-sql so that readyset itself can respond to the
DEALLOCATE request (coming in a followup commit), rather than the
current behavior which is to solely proxy to the upstream.

Fixes: REA-3983
Change-Id: I22972e8d3295236149ad9658f6a6aecb2ad9adb2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6719
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>