server: Log if a parse round trip on a new cache fails
If we are able to parse a select statement but fail to perform a
display().parse() round trip, we will fail to load a stored 'create
cache' statement after a backwards incompatible upgrade.

This commit adds a check and logs an error if we detect we will fail
this, so that we can fix the issue before upgrading.

Change-Id: Ied2367109b667f69cdbf963eb1ff6d15ec3b22b4