server: Store original 'create cache' strings in authority
Rather than parsing and displaying `create cache` statements, for
explicit `create cache` statements, take some extra time to clone and
pass along the full, original query string and store that instead.

The rationale for doing this is: if we succeeded in parsing the query
originally, we are guaranteed to succeed in doing it again. But if we
parse and then display and re-parse, that opens us up to having a bug in
our display implementation.

Change-Id: I7c77bc35f52ac7006f9ef9f59e04393a9664ec07
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6274
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI