adapter: Avoid deadlock for evicted statuses
If the lru cache evicts from the statuses in the persistent handle of
the query status cache, we may get into a situation where
`with_mut_status`, which takes an arbitrary function, can try to write
the RwLock guarding the statuses. This was the case before this commit,
which resulted in a deadlock when that happened.

This commit removes the existing RwLock.write() calls within
`with_mut_status` to avoid the deadlock and also changes the persistent
handle's `insert_with_status` to time out if it cannot acquire a write
lock, preventing deadlocks and writing a warning log if the code changes
in the future to re-introduce a double lock here.

Release-Note-Core: Fixed a race condition that could result in a
  deadlock.

Change-Id: I014d89aebb8fcbd996200976a6047a4b65440102
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6280
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>