readyset-client: Implement Change::from_str
This allows us to contruct a single `Change` from a ddl query string,
similarly to how we already support creating a ChangeList from a query
string containing multiple ddl statements.

Creating a single Change will be useful for upgrades when we need to
associate a particular change with a schema_search_path, which is added
at the ChangeList abstraction level.

This implementation does not yet implement from_str for DropView and
DropTable changes , because those do not have a 1:1 query:change
mapping.

Change-Id: I2b774dcff30b9fab1ee75557296c8bda19b19e76