controller: Load create cache statements from authority
If a new leader tries to deserialize the controller state and fails, we
need to re-create the caches we have saved off to the side. The first
step to doing this is loading them from the authority and passing them
up so that we can re-run them once snapshotting is done.

Change-Id: Ie8ebf793502d3c48b84a40fc329b44a11d845361
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6303
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>