adapter: Extract match from query_noria_extensions
`query_noria_extensions` had a bit of an awkward return time with an
Option<Result<Result>>, which can be cleaned up by checking that the
SqlQuery being passed into it is indeed a noria(readyset) extension
before calling the function. This also allows more ergonomic use of '?'
inside of `query_noria_extensions`.

Change-Id: I0b5a0b75cf9ce60a0fc919839ef69e8e146a6026
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6311
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>