treewide: Store 'create cache' statements in adapter
Now that we are string the original, unparsed 'create cache' statement,
we don't need to store it from the controller, but can do it earlier on,
in the adapter.

This change refactors things for this earlier storage, removing
RecipeSpec and the related plumbing needed to do it after an
`extend_recipe` finishes.

We will need some extra logic to handle the case where an `extend
recipe` fails after we store a `create cache` statement, which will come
in a separate commit.

Change-Id: Ibdfcf95c004a1673d8e6c6277eeedfc280c4bd86