Rename 'CreateCacheRequest' to 'CacheDDLRequest'
We will be storing `drop cache` statements along with `create cache`
ones, so rename the struct storing them to be more generic.

We can't (at least trivially) remove a `create cache request` when
processing a `create cache request` because we don't know what the id
will be after rewriting etc.

Change-Id: I29b0eef95fde802af2413f9d46dacbfbc98625b5