Add "Drop cache" statements to the authority
In order to restore the correct set of caches after a backwards
incompatible upgrade, we need to re-process or coalesce "create cache"
and "drop cache" statements. Coalescing them is not trivial because
`drop cache` uses the query id OR an alias of the cache. So store drop
cache requests alongside the create cache requests so that we can replay
them on the upgrades.

Change-Id: Ia3eab854f9155d1bd7b76220fe7f2c0fa29ec9ff