server: Restore caches after incompat. upgrade
If we have detected a backwards incompatible upgrade, load the stored
cache ddl requests (create cache and drop cache statements) from the
authority, and then re-run them to restore the state of the dataflow
graph before the upgrade.

This commit handles and tests intermingled `create cache` and `drop
cache` statements, but does not yet fully handle error conditions that
may be encountered in `extend_recipe`, which will come in a separate
commit.

Change-Id: I60af8a6112a112edda5c994ffdccaa24837b0184
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6321
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>