treewide: Modernize clap usage
Using `#[clap...]` for all annotations has been deprecated since
version 4.0, and we are on version 4.3 (and newer versions are
out now). Using a deprecated feature isn't necessarily bad, but the
documentation for 4.0+ completely hides any mention of the
original `#[clap...]`, and trying to make any changes to our use of
clap was very confusing. Thus, all uses of `#[clap...]` have been
updated to the modern equivalents, which basically amounts to:

- structs, and nested structs, get the `#[command...]` annotation
- fields within a struct become `#[arg...]`
- enums get `#[value...]` at the outer level, not on each variant

Change-Id: I886c3c1fe512feca8a1d7e561ba40c7e2d7cf4d6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6331
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>