replicators: Fix repl slot exists check
Previously, we had hard-coded the name "readyset" in our check to see if
a replication slot already exists for the given deployment upon
replicator startup. This check won't work if the name of the slot was
customized via the "--replication-server-id" flag. This commit removes
the hard-coded string in favor of the variable that would contain the
customized name.

Release-Note-Core: Fixed an issue where ReadySet failed to start up if
  the instance was passed a "--replication-server-id"
Change-Id: I666acab6e723c2f3bd4ee4379570147ef700c185