treewide: bump nightly rust version to 2023-11-09
Bump nightly rust version to 2023-11-09. Most changes, other than
simple formatting updates, include:

- change `try_into().unwrap()` -> `into()`
- change `or_insert_with(<Type>::new)` -> `or_default()`
- change `vec.get(0)` -> `vec.first()`
- `DialectDisplay::display()` - doesn't need declare the `Copy`
trait on return type any longer
- `return_position_impl_trait_in_trait` - This feature was stabilized
in rust 1.75.0, and thus removed
- A handful of structs have custom `PartialOrd` impls that are
slightly non-conforming, yet still adhere to the rule that they are
compatibile with `Ord` impls. However, this is enough to trigger the
ire of clippy. These structs include
`nom_sql::{Float,Double,SqlIdentifier,sql_type::EnumVariants},
and `OutputColumn` in readyset-server. Those impls get the
`#[allow(clippy::non_canonical_partial_ord_impl)]` tag.
- An experimental feature of collections called `drain_filter` was
replaced with `extract_if` (in feature as well as method names).
The behavior changed, as well: previously, if the returned iterator
was not explicitly consumed, it would be effectively consumed on drop.
Under `extract_if`, however, the iterator is not explicitly consumed
on drop, and thus we need to explicitly consume it on the few code
paths that relied on the drop behavior.

Change-Id: I2508f2e76a2a5a18f0c0d2b98b7301980d0ba620
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6349
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>