readyset-client: Add schema offset to AuthorityControl
This adds a way to read the schema replication offset that we are now
storing off to the side when updating the controller state. It can be
used to restore the schema replication offset if we fail to load the
ControllerState due to a backwards incompatible change in
ControllerState serialization format.

Change-Id: Idfdc6835b41a06e18f84196c3de64302997f9154
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6357
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>