test-helpers: Set durability mode explicitly
Before this commit, the TestBuilder had a `fn persistent` that used
DeleteOnExit persistence mode--it writes to disk, but then throws it
away and doesn't have a mechanism for re-using it.

This changes the persistence configuration to be a 1:1 mapping to
`DurabilityMode`, defaulting to MemoryOnly. If `PErmanent` is used, the
caller is responsible for cleaning up the tempdir (it will be
automatically cleaned up when it goes out of scope, so this is fairly
easy to do as the caller).

Change-Id: I822d82e902d7d3e7addd164a2d6281ba25e06c20